(PC-16806) fix(algolia): move mockedAlgoliaResponse and mockedVenue from mockedResponses to __mocks__ folder
