+++ src/features/home/components/tests/OffersModule.native.test.tsx
@@ -7,7 +7,7 @@ import {
   DisplayParametersFields,
   ContentTypes,
 } from 'features/home/contentful/contentful'
-import { mockedAlgoliaResponse } from 'libs/algolia/mockedResponses/mockedAlgoliaResponse'
+import { mockedAlgoliaResponse } from 'libs/algolia/__mocks__/mockedAlgoliaResponse'
 import { analytics } from 'libs/firebase/analytics'
 import { SearchHit, transformHit } from 'libs/search'
 import { flushAllPromises, act, fireEvent, render } from 'tests/utils'
