+++ src/features/home/components/tests/RecommendationModule.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { RecommendationModule } from 'features/home/components/RecommendationModule'
 import { ContentTypes, DisplayParametersFields } from 'features/home/contentful'
-import { mockedAlgoliaResponse } from 'libs/algolia/mockedResponses/mockedAlgoliaResponse'
+import { mockedAlgoliaResponse } from 'libs/algolia/__mocks__/mockedAlgoliaResponse'
 import { analytics } from 'libs/firebase/analytics'
 import { render } from 'tests/utils'
